+++ cluster-cm/src/main/java/com/sequenceiq/cloudbreak/cm/polling/ClouderaManagerSyncCommandPollerObject.java
@@ -1,7 +1,5 @@
 package com.sequenceiq.cloudbreak.cm.polling;
 
-import java.math.BigDecimal;
-
 import com.cloudera.api.swagger.client.ApiClient;
 import com.sequenceiq.cloudbreak.domain.stack.Stack;
 
@@ -9,7 +7,7 @@ public class ClouderaManagerSyncCommandPollerObject extends ClouderaManagerComma
 
     private final String commandName;
 
-    public ClouderaManagerSyncCommandPollerObject(Stack stack, ApiClient apiClient, BigDecimal id, String commandName) {
+    public ClouderaManagerSyncCommandPollerObject(Stack stack, ApiClient apiClient, Integer id, String commandName) {
         super(stack, apiClient, id);
         this.commandName = commandName;
     }
